refactor code to introduce new forms of chaos
