+++ api/src/main/java/org/arquillian/cube/q/api/Q.java
@@ -4,30 +4,11 @@ import java.util.concurrent.TimeUnit;
 
 public interface Q {
 
-    Action on(String machine, int port);
-    
-    public interface Action {
+    void exec() throws Exception;
+    void exec(Perform perform) throws Exception;
+    void exec(RunCondition runCondition, Perform perform) throws Exception;
 
-        Action down();
-        Action timeout(TimeoutType timeType);
-        Action timeout(TimeoutType timeType, ToxicityType toxicityType, ToxicDirectionStream toxicDirectionStream);
-        Action latency(LatencyType latencyType);
-        Action latency(LatencyType latencyType, JitterType jitterType, ToxicityType toxicityType, ToxicDirectionStream toxicDirectionStream);
-        Action bandwidth(RateType rateType);
-        Action bandwidth(RateType rateType, ToxicityType toxicityType, ToxicDirectionStream toxicDirectionStream);
-        Action slowClose(DelayType delayType);
-        Action slowClose(DelayType delayType, ToxicityType toxicityType, ToxicDirectionStream toxicDirectionStream);
-        Action slice(SliceAverageSizeType sliceAverageSizeType, DelayType delayType);
-        Action slice(SliceAverageSizeType sliceAverageSizeType, DelayType delayType, SliceSizeVariationType sliceSizeVariationType, ToxicityType toxicityType, ToxicDirectionStream toxicDirectionStream);
-
-        void exec() throws Exception;
-        void exec(Perform perform) throws Exception;
-        void exec(RunCondition runCondition, Perform perform) throws Exception;
-
-    }
-    
     public interface Perform {
-        
         void execute() throws Exception;
     }
 
@@ -35,11 +16,7 @@ public interface Q {
         boolean isExecutable();
     }
 
-    public static enum ToxicDirectionStream {
-        DOWNSTREAM, UPSTREAM
-    }
-
-    public static class DurationRunCondition implements RunCondition {
+    public static class DurationRunCondition implements Q.RunCondition {
 
         protected long finishTime;
 
@@ -58,7 +35,7 @@ public interface Q {
         }
     }
 
-    public static class IterationRunCondition implements RunCondition {
+    public static class IterationRunCondition implements Q.RunCondition {
         protected final long iterations;
         protected long currentIteration = 0;
 
@@ -79,117 +56,6 @@ public interface Q {
             } else {
                 return false;
             }
-        }
-    }
-
-    public static final class ToxicityType extends FloatType {
-
-        protected ToxicityType(float value) {
-            super(value);
-        }
-
-        public static ToxicityType fullToxic() {
-            return new ToxicityType(1f);
-        }
-
-        public static ToxicityType noToxic() {
-            return new ToxicityType(0f);
-        }
-
-        /**
-         * Toxicity value. It is a percentage between 0 and 1.
-         * @param toxicity
-         * @return
-         */
-        public static ToxicityType toxicity(float toxicity) {
-            return new ToxicityType(toxicity);
-        }
-    }
-
-    public static final class SliceSizeVariationType extends LongType {
-
-        protected SliceSizeVariationType(long value) {
-            super(value);
-        }
-
-        public static SliceSizeVariationType sliceSizeVariation(long sizeVariation) {
-            return new SliceSizeVariationType(sizeVariation);
-        }
-    }
-
-    public static final class SliceAverageSizeType extends LongType {
-
-        protected SliceAverageSizeType(long value) {
-            super(value);
-        }
-
-        public static SliceAverageSizeType sliceAverageSize(long average) {
-            return new SliceAverageSizeType(average);
-        }
-    }
-
-    public static final class DelayType extends LongType {
-
-        protected DelayType(long value) {
-            super(value);
-        }
-
-        public static DelayType delay(long delay) {
-            return new DelayType(delay);
-        }
-
-        public static DelayType delay(long delay, TimeUnit timeUnit) {
-            return new DelayType(timeUnit.toMillis(delay));
-        }
-    }
-
-    public static final class RateType extends LongType {
-
-        protected RateType(long value) {
-            super(value);
-        }
-
-        public static RateType rate(long rate) {
-            return new RateType(rate);
-        }
-    }
-
-    public static final class JitterType extends IntegerType {
-
-        protected JitterType(int value) {
-            super(value);
-        }
-
-        public static JitterType jitter(int jitter) {
-            return new JitterType(jitter);
-        }
-    }
-
-    public static final class LatencyType extends LongType {
-        protected LatencyType(long value) {
-            super(value);
-        }
-
-        public static LatencyType latency(long time) {
-            return new LatencyType(time);
-        }
-
-        public static LatencyType latency(long time, TimeUnit timeUnit) {
-            return new LatencyType(timeUnit.toMillis(time));
-        }
-    }
-
-    public static final class TimeoutType extends LongType {
-        protected TimeoutType(long value) {
-            super(value);
-        }
-
-        public static TimeoutType timeout(long time) {
-            return new TimeoutType(time);
-        }
-
-        public static TimeoutType timeout(long time, TimeUnit timeUnit) {
-            return new TimeoutType(timeUnit.toMillis(time));
         }
     }
