+++ toxic/src/main/java/org/arquillian/cube/q/toxic/QNetworkChaosResourceProvider.java
@@ -1,24 +1,25 @@
-package org.arquillian.cube.q.core;
+package org.arquillian.cube.q.toxic;
 
 import java.lang.annotation.Annotation;
 
+import org.arquillian.cube.q.api.NetworkChaos;
 import org.arquillian.cube.q.api.Q;
 import org.jboss.arquillian.core.api.Instance;
 import org.jboss.arquillian.core.api.annotation.Inject;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.arquillian.test.spi.enricher.resource.ResourceProvider;
 
-public class QResourceProvider implements ResourceProvider {
+public class QNetworkChaosResourceProvider implements ResourceProvider {
 
     @Inject
-    private Instance<Q> qInst;
+    private Instance<NetworkChaos> networkChaosInst;
     
     public boolean canProvide(Class<?> type) {
-        return type.isAssignableFrom(Q.class);
+        return type.isAssignableFrom(NetworkChaos.class);
     }
 
     public Object lookup(ArquillianResource resource, Annotation... qualifiers) {
-        return qInst.get();
+        return networkChaosInst.get();
     }
 
 }
