+++ toxic/src/main/java/org/arquillian/cube/q/toxic/QNetworkChaosToxicCreator.java
@@ -1,5 +1,6 @@
 package org.arquillian.cube.q.toxic;
 
+import org.arquillian.cube.q.api.NetworkChaos;
 import org.arquillian.cube.q.api.Q;
 import org.arquillian.cube.q.toxic.client.ToxiProxyScenario;
 import org.jboss.arquillian.core.api.InstanceProducer;
@@ -7,12 +8,12 @@ import org.jboss.arquillian.core.api.annotation.ApplicationScoped;
 import org.jboss.arquillian.core.api.annotation.Inject;
 import org.jboss.arquillian.core.api.annotation.Observes;
 
-public class QToxicCreator {
+public class QNetworkChaosToxicCreator {
 
     @Inject @ApplicationScoped
-    private InstanceProducer<Q> qInst;
+    private InstanceProducer<NetworkChaos> qInst;
     
     public void createProxy(@Observes ToxiProxyScenario proxy) {
-        qInst.set(new QToxic(proxy));
+        qInst.set(new QNetworkChaosToxic(proxy));
     } 
 }
