+++ toxic/src/main/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,1 @@
-org.arquillian.cube.q.toxic.QToxicExtension
+org.arquillian.cube.q.toxic.QNetworkChaosToxicExtension
